http://bugzilla.zimbra.com/show_bug.cgi?id=6219 bug: 6219 No more menu on upgrade, unless there are unconfigured items...
Copied from Perforce
 Change: 22195

+++ rpmconf/Install/Util/utilfunc.sh
@@ -507,7 +507,7 @@ restoreCerts() {
 	if [ -f "$SAVEDIR/cacerts" ]; then
 		cp $SAVEDIR/cacerts /opt/zimbra/java/jre/lib/security/cacerts
 	fi
-	if [ -f "$SAVEDIR/keystore" ]; then
+	if [ -f "$SAVEDIR/keystore" -a -d "/opt/zimbra/tomcat/conf" ]; then
 		cp $SAVEDIR/keystore /opt/zimbra/tomcat/conf/keystore
 	fi
 	if [ -f "$SAVEDIR/smtpd.key" ]; then
